LayoutTests/fast/events/ios/did-not-handle-meaningful-click.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no">
<style>
body, html {
    font-size: 18px;
    font-family: system-ui;
    width: 100%;
    height: 100%;
    margin: 0;
}
</style>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("This test is intended to exercise the 'meaningful click' heuristic, and requires WebKitTestRunner.");

    let text = document.getElementById("text");
    let button = document.getElementById("target");

    async function simulateTapAndCheckDidHandleTapAsMeaningfulClick(element, expectation) {
        didHandleTap = false;
        didHandleTapAsMeaningfulClick = false;
        testRunner.installDidHandleTapCallback(wasMeaningful => {
            didHandleTap = true;
            didHandleTapAsMeaningfulClick = wasMeaningful;
        });

        await UIHelper.activateElement(element);
        await UIHelper.waitForDoubleTapDelay();
        shouldBeTrue("didHandleTap");
        if (expectation)
            shouldBeTrue("didHandleTapAsMeaningfulClick");
        else
            shouldBeFalse("didHandleTapAsMeaningfulClick");

        testRunner.clearTestRunnerCallbacks();
    }

    await simulateTapAndCheckDidHandleTapAsMeaningfulClick(text, false);
    await simulateTapAndCheckDidHandleTapAsMeaningfulClick(button, true);

    document.body.addEventListener("click", () => { });
    await simulateTapAndCheckDidHandleTapAsMeaningfulClick(text, false);

    document.body.addEventListener("mousedown", (event) => {
        event.preventDefault();
    });
    await simulateTapAndCheckDidHandleTapAsMeaningfulClick(text, true);

    text.remove();
    button.remove();

    finishJSTest();
});
</script>
</head>
<body>
    <button id="target" onclick="testPassed('clicked on button'); event.stopPropagation();">Hello world</button>
    <p id="text">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Aenean laoreet nisi felis, eu scelerisque dolor imperdiet in. Sed at porttitor purus, in placerat justo. Mauris at augue ac purus viverra molestie. Fusce ultrices ante at pellentesque lobortis. Sed pellentesque, ipsum eu semper mollis, arcu lectus dignissim enim, in maximus tortor ligula id ipsum. Nulla dignissim diam a mi tempor porta. Etiam tempor ac augue non dapibus. Donec euismod vel turpis in ultrices. Ut accumsan ultrices arcu vitae venenatis. In urna velit, accumsan sit amet arcu vel, vestibulum pulvinar leo. Aliquam vulputate euismod ultrices. Sed congue vestibulum libero quis ornare.</p>
    <div id="description"></div>
    <div id="console"></div>
</body>
</html>
